requirements.txt
# Core dependencies
aiohttp==3.9.1
asyncio==3.4.3
numpy==1.26.2
cryptography==41.0.7
python-dotenv==1.0.0
PyJWT==2.8.0

# Solana libraries (hypothetical - these would be needed for real implementation)
# solana==0.30.2
# solders==0.18.1
# base58==2.1.1

# Optional monitoring and logging
colorlog==6.8.0
prometheus-client==0.19.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
mypy==1.7.1